Move list jobs api call struct to struct/syncserver
